Replace decorators with prependers
It fixes Zeitwerk errors (thus Rails 6 errors) due to decorators not
declaring constants that Zeitwerk infers by their paths.
