+++ lib/solidus_sale_prices.rb
@@ -2,5 +2,7 @@ require 'spree_core'
 
 require 'solidus_support'
 
+require 'prependers'
+
 require 'solidus_sale_prices/engine'
 require 'deface'
