+++ lib/solidus_sale_prices/factories.rb
@@ -1,3 +1,5 @@
+module SolidusSalePrices::Factories; end
+
 FactoryBot.define do
   # Define your Spree extensions Factories within this file to enable applications, and other extensions to use and override them.
   #
