+++ solidus_sale_prices.gemspec
@@ -22,6 +22,7 @@ Gem::Specification.new do |s|
 
   s.add_runtime_dependency 'deface', '~> 1.0'
 
+  s.add_dependency 'prependers'
   s.add_dependency "solidus_api", solidus_version
   s.add_dependency "solidus_backend", solidus_version
   s.add_dependency "solidus_core", solidus_version
